Fix metric loading for AWS/ES
AWS ElasticSearch does not return any metrics when the client ID and the cluster are not set. This PR takes the `awsDimensionSelect` and puts that into the dimension filters with the values, so the metrics will actually return a value.

This is slightly brittle, but so is the AWS API apparently

Signed-off-by: Alex Nederlof <alex@nederlof.com>