Fix Docker builds on linux/arm64
Follow-up to #372, #374, #363.

Update the base images used in docker builds. Install Maven from the
tarball, apt pulls in the wrong JDK as a dependency.

Set up the context and builder for buildx.

Work around a bug in Java 13+ on linux/arm64 by using the older "vfork"
method. Update the surefire plugin.

Always run the image build, even for PRs, so that we can catch breakage
here before it hits the main branch.

Signed-off-by: Matthias Rampke <matthias@prometheus.io>